restructured architectures
